Sample data that can be analysed using the Dynameta shiny app

An example dataset containing data collected for a GLiTRS meta-analysis investigating the effect of insecticide application on Odonata abundance. This can be analysed using the Dynameta shiny app.

Usage

data(sample_data)

Format

A data frame with 47 rows and 46 variables:

Paper_ID

The paper from which the comparison was drawn. (character)

Included

The exclusion or inclusion status of the paper. (character)

Observation_ID

The comparison of that row, which should be unique to each row. (character)

Author

The full list of authors of the paper. (character)

Year

The year of publication of the paper. (integer)

Title

The paper title. (character)

Journal

The journal the paper was published in. (character)

DOI

The DOI of the paper. (character)

URL

The URL of the paper. (character)

Language

The language of the paper. (character)

Database

The database from which the paper was found. (character)

Latitude

The latitude coordinate of the observation in that row. (double)

Longitude

The longitude coordinate of the observation in that row. (double)

Country

The country in which the observation in that row is located. (character)

Taxa_level

The taxonomic level that is common to all of the biodiversity recorded in that row. (character)

Taxa_name

The name of the taxonomic level that is common to all of the biodiversity recorded in that row. (character)

Order

The taxonomic order of the biodiversity measured in that row, for both the treatment and control. (character)

Family

The taxonomic family of the biodiversity measured in that row, for both the treatment and control. (character)

Genus

The taxonomic genus of the biodiversity measured in that row, for both the treatment and control. (character)

Binomial

The Latin binomial of the species in that row, if it has been recorded. (character)

Life_history_stage

The life-history stage of the taxa in that row. (character)

Experimental_year_start

The year in which data collection started. (integer)

Experimental_year_end

The year in which data collection ended. (integer)

Sampling_method

The sampling method used to sample the biodiversity measure. (character)

Biodiversity_metric

The metric of biodiversity measured. Biodiversity_metric should typically be one of “Abundance”, “Richness”, or “Biomass”. (character)

Unit

The more specific unit of measure for the biodiversity recorded in that row. (character)

IUCN_threat_category_1

Broadest level of IUCN threat. (character)

IUCN_threat_category_2

Secondary level of IUCN threat. (character)

IUCN_threat_category_3

Tertiary level of IUCN threat. (character)

Treatment

Name of the threat measured in that row (e.g. insecticide). (character)

Treatment_quantity

The quantity of treatment reported by the authors for that observation. (double)

Treatment_quantity_unit

The unit of measure for the numeric value in Treatment_quantity. (character)

Control

Name of the control measured in that row (e.g. no insecticide). (character)

Extracted_from

The figure or table from which the data in that row was extracted. (character)

Evidence_type

The type of evidence for the data in that row. Type of evidence should be either “Experimental” or “Quasi-experimental”. (character)

Treatment_N

The number of treatment sites from which the mean and error values were drawn. (integer)

Treatment_Mean

The mean biodiversity value across all treatment sites for that comparison. (double)

Treatment_error

The raw treatment error value reported by the authors for that observation. (double)

Treatment_error_type

The type of error recorded in Treatment_error. Typically this will be one of Standard error, Standard deviation, or 95% confidence interval. (character)

Control_N

The number of control sites from which the mean and error values were drawn. (integer)

Control_Mean

The mean biodiversity value across all control sites for that comparison. (double)

Control_error

The raw control error value reported by the authors for that observation. (double)

Control_error_type

The type of error recorded on Control_error. Typically this will be one of “Standard error”, “Standard deviation”, or “95% confidence interval”. (character)

Contributor_name

Name of the contributor for that row. (character)

Search_date

Date on which the final optimised search is carried out. (date)

Notes

This field should be used for any relevant descriptions that don’t fall within core data, can’t be categorised into clearly defined factors in the associated metadata, or provide explanations for the associated metadata. (character)